scenario_service/requirements.txt
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
kubernetes>=30.1.0
python-multipart>=0.0.9
pyyaml>=6.0.2
gunicorn>=23.0.0,<24